RAG/backend/requirements.txt
# backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pymongo>=4.6.0
google-generativeai>=0.3.0
python-multipart>=0.0.7